Move group management to users view
